:bug: :art: fix:細かいバグ修正、リファクタリング
