+++ root/components/Footer.tsx
@@ -19,15 +19,7 @@ export const Footer = () => {
   return (
     <FooterContainer>
       <Typography variant="h6" align="center" gutterBottom>
-        Footer
-      </Typography>
-      <Typography
-        variant="subtitle1"
-        align="center"
-        color="textSecondary"
-        component="p"
-      >
-        Something here to give the footer a purpose!
+        BlogFavo
       </Typography>
       <Copyright />
     </FooterContainer>
